Clean code: Rename test case StopArrivalsTestData to BasicItineraryTestCase.
